prueba_back/requirements.txt
Django==2.1.5
django-cors-headers==2.1.0
djangorestframework==3.7.1
psycopg2==2.7.7
requests==2.18.4